+++ test/filters.test.ts
@@ -12,11 +12,16 @@ test('Get filtered rows of dataset with simple "and" filter', async () => {
     const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'AGE', operator: 'gt', value: 13 },
-            { variable: 'SEX', operator: 'eq', value: 'M' }
+            { variable: 'SEX', operator: 'eq', value: 'M' },
         ],
-        connectors: ['and']
+        connectors: ['and'],
     });
-    const rows = await data.getData({ start: 0, length: 5, filter, filterColumns: ['NAME', 'SEX', 'AGE'] });
+    const rows = await data.getData({
+        start: 0,
+        length: 5,
+        filter,
+        filterColumns: ['NAME', 'SEX', 'AGE'],
+    });
     expect(rows.length).toBeLessThanOrEqual(5);
     expect(rows).toMatchSnapshot();
 });
@@ -30,10 +35,36 @@ test('Get filtered rows of dataset with simple "or" filter', async () => {
     const filter = new Filter('dataset-json1.1', metadata.columns, {
         conditions: [
             { variable: 'AGE', operator: 'gt', value: 12 },
-            { variable: 'SEX', operator: 'eq', value: 'M' }
+            { variable: 'SEX', operator: 'eq', value: 'M' },
         ],
-        connectors: ['or']
+        connectors: ['or'],
     });
-    const rows = await data.getData({ start: 0, filter, filterColumns: ['NAME', 'SEX', 'AGE'] });
+    const rows = await data.getData({
+        start: 0,
+        filter,
+        filterColumns: ['NAME', 'SEX', 'AGE'],
+    });
     expect(rows.length).toEqual(16);
+});
+
+test('Get filtered rows of dataset with dynamic length', async () => {
+    const filePath = path.join(__dirname, '/data/sample.sas7bdat');
+
+    const data = new DatasetSas7BDat(filePath);
+
+    const metadata = await data.getMetadata();
+
+    const filter = new Filter('dataset-json1.1', metadata.columns, {
+        conditions: [{ variable: 'SEX', operator: 'eq', value: 'M' }],
+        connectors: [],
+    });
+    const rows = await data.getData({
+        start: 0,
+        length: 5,
+        filter,
+        filterColumns: ['NAME', 'SEX', 'AGE'],
+        dynamicLength: true,
+    });
+    expect(rows.length).toBeLessThanOrEqual(5);
+    expect(rows).toMatchSnapshot();
 });