+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [0.1.4]
+* Fixing creation/modification time
+* Adding dynamic length flag for reading data with filters
+
+## [0.1.3]
+* Fixing an issue when filtering datasets with the number of rows > page length
+* Updating compilation instructions for Windows
+
 ## [0.1.2]
 * Adding pre-builds